Improved ability to close pop up flashcard settings menu and added css
Added a button next to menu header that will close window

removed window event listener that checked for a click outside of menu and instead added an onClick on the "darken-screen" div that covers background of pop up menu. This made the it more consistent and easier to work with.

Added styling for the popup menu to make it more prominent, centered and to make it more easier for the user to understand how to close it